+++ src/rgb/s_rgb/mod.rs
@@ -38,6 +38,15 @@ impl<T: Component> SRgb<T> {
 		Self(colour)
 	}
 
+	/// Reinterprets a raw RGB colour as sRGB.
+	///
+	/// The provided colour is *scaled* to fit the sRGB gamut.
+	#[inline(always)]
+	#[must_use]
+	pub const fn from_rgb(colour: Rgb<T>) -> Self {
+		Self(colour)
+	}
+
 	/// Maps the sRGB colour's channels.
 	#[inline]
 	#[must_use]
@@ -74,21 +83,56 @@ macro_rules! impl_conversions {
 	($($tys:ty),+$(,)?) => {
 		$(
 			impl ::polywave::rgb::SRgb<$tys> {
+				/// Transfers a raw RGB value to perceptual RGB, scaling to the sRGB gamut.
+				///
+				/// The transfer is done as by the sRGB transfer function.
+				#[cfg(feature = "std")]
+				#[inline]
+				#[must_use]
+				pub fn transfer(colour: Rgb<$tys>) -> Self {
+					let colour = colour.map(|mut colour| {
+						let sign = colour;
+
+						colour = colour.abs();
+
+						colour = if colour > 0.003_130_800 {
+							colour.powf(const { 1.0 / 2.4 }) * 1.055 - 0.055
+						} else {
+							colour * 12.920
+						};
+
+						colour = colour.copysign(sign);
+
+						colour
+					});
+
+					Self::from_rgb(colour)
+				}
+
 				/// "Untransfers" the gamma-encoded sRGB.
 				///
-				/// sRGB channels are encoded using the *transfer* function.
-				/// This method is the inverse of this gamma function.
+				/// sRGB channels are encoded using the [*transfer*](https://en.wikipedia.org/wiki/SRGB#Transfer_function_(%22gamma%22)) function (see [`transfer`](Self::transfer)).
+				/// This method serves as the inverse of this function.
 				///
 				/// Note that the returned value is no longer sRGB as sRGB is strictly gamma-encoded.
 				#[cfg(feature = "std")]
+				#[inline]
 				#[must_use]
 				pub fn untransfer(self) -> Rgb<$tys> {
-					self.as_rgb().map(|colour| {
-						if colour > 0.040_450 {
+					self.as_rgb().map(|mut colour| {
+						let sign = colour;
+
+						colour = colour.abs();
+
+						colour = if colour > 0.040_450 {
 							((colour + 0.055) / 1.055).powf(2.4)
 						} else {
 							colour / 12.920
-						}
+						};
+
+						colour = colour.copysign(sign);
+
+						colour
 					})
 				}
 			}